% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findRscript}
\alias{findRscript}
\title{Returns list of R installation paths.}
\usage{
findRscript(path, pattern)
}
\arguments{
\item{path}{Character vector of search paths. Passed to \code{\link{list.files}}}

\item{pattern}{regular expression.  Passed to \code{\link{list.files}}}
}
\description{
Given a vector of search paths, returns all instances of R installations.
}
\details{
This function is a wrapper around \code{list.files} setting the search pattern to \code{Rscript.exe}.
}

