% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/urbanekPlot.R
\name{urbanek2df}
\alias{urbanek2df}
\title{Converts Urbanek tests to data frame.}
\usage{
urbanek2df(x)
}
\arguments{
\item{x}{The result of a call to \code{\link{urbanek2.5}}}
}
\description{
Converts Urbanek tests to data frame.
}

