% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/urbanekPlot.R
\name{urbanekCombine}
\alias{urbanekCombine}
\title{Combines results from Urbanek tests}
\usage{
urbanekCombine(x)
}
\arguments{
\item{x}{The result of a call to \code{\link{urbanek2.5}}}
}
\description{
Combines results from Urbanek tests
}

