% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/urbanekPlot.R
\name{urbanekPerformance}
\alias{urbanekPerformance}
\title{Computes urbanek performance improvement.}
\usage{
urbanekPerformance(x)
}
\arguments{
\item{x}{The result of a call to \code{\link{urbanek2.5}}}
}
\description{
Computes urbanek performance improvement.
}

