% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/urbanekPlot.R
\name{urbanekPlot}
\alias{urbanekPlot}
\title{Create plot of urbanek test results.}
\usage{
urbanekPlot(x)
}
\arguments{
\item{x}{The result of a call to \code{\link{urbanek2.5}}}
}
\description{
Create plot of urbanek test results.
}

