% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/version.time.R
\name{version.time}
\alias{version.time}
\title{Run script in different versions of R and capture system time.}
\usage{
version.time(expr, rVersions)
}
\arguments{
\item{expr}{Expression}

\item{rVersions}{Character vector containing path to Rscript in different installations of R}
}
\description{
Run script in different versions of R and capture system time.
}
\examples{
# Find installed versions of Rscript

rversion <- paste(R.version$major, R.version$minor, sep = ".")
rscript <- findRscript(version = rversion)


# Configure which installed version to use

# Compute vector mean in different R installations

version.time({
  foo <- rnorm(1e6)
  mean(foo)
} , rscript)


# Compute matrix cross product in different R installations

version.time({
  m <- matrix(runif(100), nrow=10)
  crossprod(m)
} , rscript)
}

