% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevoBenchmark-methods.R
\name{urbanekPerformance}
\alias{urbanekPerformance}
\title{Computes urbanek performance improvement.}
\usage{
urbanekPerformance(x)
}
\arguments{
\item{x}{RevoBenchmark object}
}
\description{
Computes urbanek performance improvement.
}

