% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.time.R
\name{version.time}
\alias{version.time}
\title{Run script in different versions of R and capture system time.}
\usage{
version.time(rVersions, expr, file)
}
\arguments{
\item{rVersions}{Character vector containing path to Rscript in different installations of R}

\item{expr}{Expression to evaluate}

\item{file}{If specified, evaluates the code in the file, rather than \code{expr}}
}
\description{
Run script in different versions of R and capture system time.
}
\examples{
# Find installed versions of Rscript

rversion <- paste(R.version$major, R.version$minor, sep = ".")
rscript <- findRscript(version = rversion)

# Configure which installed version to use

# Compute vector mean in different R installations

version.time(rscript, {
  set.seed(1)
  foo <- rnorm(1e4)
  mean(foo)
})


# Compute matrix cross product in different R installations

version.time(rscript, {
  set.seed(42)
  m <- matrix(runif(100), nrow=10)
  crossprod(m)
})

}

