% Generated by roxygen2 (4.0.2): do not edit by hand
\name{version.time}
\alias{version.time}
\title{Run script in different versions of R and capture system time.}
\usage{
version.time(expr, rVersions)
}
\arguments{
\item{expr}{Expression}

\item{rVersions}{Character vector containing path to Rscript in different installations of R}
}
\description{
Run script in different versions of R and capture system time.
}

